refactor : Graph의 가중치를 결정하는 로직을 EdgeWeightType으로 이동

+++ src/main/java/wooteco/subway/admin/domain/GraphService.java
@@ -39,16 +39,16 @@ public class GraphService {
 	}
 
 	private int sumDistance(GraphPath<Long, CustomEdge> path) {
-		return path.getEdgeList()
+		return (int)path.getEdgeList()
 			.stream()
-			.mapToInt(CustomEdge::getDistance)
+			.mapToDouble(CustomEdge::getDistance)
 			.sum();
 	}
 
 	private int sumDuration(GraphPath<Long, CustomEdge> path) {
-		return path.getEdgeList()
+		return (int)path.getEdgeList()
 			.stream()
-			.mapToInt(CustomEdge::getDuration)
+			.mapToDouble(CustomEdge::getDuration)
 			.sum();
 	}
